Only undo if buffer is modified
Fix accidental ctrl-p branches your undo history.

Recently, I often accidentally ctrl-p in vim when I meant to do it in
another application. This :undo would revert my last good change and
invoking unite creates a new change in the undotree. Since the undo
history isn't linear, I can't redo to get back to my file's current
state. Normal usage of history_yank-cycle is to p and then C-p, so we
expect to be in the middle of editing.

Test
* save, ctrl-p, u